// src/type/price.d.ts
export interface PriceJson {
  fareRevisionStatus?: string
  toLineIndex?: string
  fromLineIndex?: string
  kind: string
  index?: string
  selected?: string
  Type?: string
  Oneway?: string
  RevisionStatus?: string
  Round?: string
  passClassIndex?: string
  Name?: string
  vehicleIndex?: string
}